Move data ID extraction utility code between subpackages.
