+++ python/lsst/daf/butler/queries/predicate_constraints_summary.py
